basic splash and calculus improved

+++ app/src/main/java/com/soces_fni/mdoser/model/AbstractOperation.java
@@ -6,5 +6,5 @@ public abstract class AbstractOperation {
      * method for calculate needed quantity.
      * @return float value.
      */
-    public abstract float CalculateQuantity();
+    public abstract double CalculateQuantity();
 }
